+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+### New features
+
+* Enable custom configuration and middlewares for the underlying HTTP client
+  ([Documentation](https://firebase-php.readthedocs.io/en/latest/setup.html#http-client-options-and-middlewares))
+
+### Bugfixes
+
 * Fixed issue when creating anonymous users: The Firebase REST API now requires empty payloads to be objects, not arrays (`{}` instead of `[]`)
 
 ## 4.10.1 - 2018-06-19